+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - PersistenceSettings.OnSaveError
 - PersistenceSettings.OnLoadError
+- LoadingUIHandler component
 
 ### Changed
 - Refactor PersistenceSettings.Save into a Task function